Deploy docs to lean-ga-docs on every build (#17)
This pushes mathlib-like docs to https://pygae.github.io/lean-ga-docs/.

Since this runs on all branches, it will push the latest of the PRs made by @pygae contributors too. I think that's ok, since we're unlikely to have many of those. We can adjust that later as we see fit.

This currently uses a forked patched version of doc-gen, but hopefully the patches will be upstreamed soon.